+++ src/main/java/com/quasiris/qsf/commons/util/ElasticUtil.java
@@ -1,5 +1,7 @@
 package com.quasiris.qsf.commons.util;
 
+import java.util.regex.Pattern;
+
 /**
  * Created by mki on 7.7.18.
  */
@@ -26,6 +28,16 @@ public class ElasticUtil {
 
     public static String value2FieldName(String value) {
         value = value.toLowerCase();
+        value = value.replaceAll(Pattern.quote("ü"), "ue");
+        value = value.replaceAll(Pattern.quote("ä"), "ae");
+        value = value.replaceAll(Pattern.quote("ö"), "oe");
+        value = value.replaceAll(Pattern.quote("ß"), "ss");
+        value = value.replaceAll(Pattern.quote("€"), "eur");
+        value = value.replaceAll(Pattern.quote("§"), "paragraph");
+        value = value.replaceAll(Pattern.quote("$"), "dollar");
+        value = value.replaceAll(Pattern.quote("@"), "at");
+        value = value.replaceAll(Pattern.quote("#"), "hash");
+        value = value.replaceAll(Pattern.quote("*"), "star");
         value = value.replaceAll("[^a-z0-9]", "_");
         return value;
     }