+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes in this project will be documented in this file.
 ## [UNRELEASED]
 ### Added
 - add embedDoc method
+- implement a value 2 filename method for elastic
 
 ## [0.2] - 2020-07-13
 ### Added